+++ src/js/looker_speedo.js
@@ -1,0 +1,14 @@
+import {makeGauge} from './main.js';
+require('../style/style.css');
+
+looker.plugins.visualizations.add({
+
+  create: function(element, config){},
+  updateAsync: function(data, element, config, queryResponse, details, doneRendering){
+    var numerator = data[0]['var1'].value
+    var denominator = data[0]['var2'].value
+    makeGauge(element, numerator, denominator)
+    doneRendering()
+
+}
+});
